Merged with my local version
The most important differences in this commit are

  (i)   A (approximately) conservative particle source
  (ii)  Support for the MUMPS linear solver
  (iii) Partially implemented avalanche source term

(i) CONSERVATIVE PARTICLE SOURCE
A big problem often manifested in NORSE simulations is that the solver
does not conserve the number of particles. This leads to a variation in,
particularly, the density through the simulation, which affects the
critical and Dreicer electric fields, and thus affect the physics. The
typical solution to this problem has previously been to increase the
momentum resolution, but due to the quadratic scaling of the non-linear
potential matrices with both nP and nXi, this quickly makes the
simulation unfeasible in both time and memory consumption. And even when
the resolution is dialed up significantly, there may still be a constant
loss of particles if a non-uniform grid is used.

With this new conservative particle source, NORSE tries to estimate in
every timestep how many particles will be added (removed) to the solution
in the next timestep, and subsequently remove (add) as many particles.

This term is easily enabled using 'NORSE.conservativeParticleSource = 1'
and is recommended for ALL simulations. It is enabled by default, and
must therefore be explicitly disabled if not desired.

(ii) SUPPORT FOR MUMPS
It was suggested by Y. Peysson that using the fast linear solver library
MUMPS for inverting the NORSE matrix could significantly improve
performance. This has previously been done to good effect in the LUKE
kinetic solver. Because of this, a MUMPS solver was implemented both as
a separate time-advance option (4), but also as part of the class
'TAInductiveE', which also solves for the electric field evolution while
solving the Fokker-Planck equation. In the end, MUMPS was found to
provide a roughly 40% improvement in performance compared to
timeAdvanceMode=0. This is still slower than the GMRES solver, but as
the iterative GMRES solver can fail in trace runaway scenarios, the
MUMPS solver may be preferred if available.

Note that MUMPS must be installed and compiled separately. Compilation
for Matlab is highly non-trivial and requires a number of special steps
(in fact, according to Y. Peysson, simply being able to call MUMPS from
Matlab was one of the major achievements in the coupling with LUKE and
MUMPS).

(iii) AVALANCHE SOURCE
Long ago, I started implementing the Chiu-Harvey avalanche source into
NORSE. Before being able to test the implementation, however, I
discovered various discrepancies between NORSE and CODE which forced me
to temporarily abandon the implementation. While the implementation here
should be fully functional, it has never been tested, and should
therefore not be used. I have merely included it here for completeness.

It is also highly unlikely that NORSE will ever be able to properly
simulate an avalanche scenario, due to the very unfavourable scalings of
the potential matrices (since a very high pMax is usually needed for
avalanche scenarios, which necessitates a large nP, the potential
matrices would likely become too large to fit into the memory of any
existing single-node computer).